refactor(frontend): extract GraphicalFields geometry into graphicalFieldsGeometry
Move the layout constants, the shared geometry types, getFitViewportPadding
and the 95-line snapToNeighbors helper out of pages/GraphicalFields.tsx
into pages/graphicalFieldsGeometry.ts. The page (1926 -> 1769 lines)
imports them back; no behavior changes.
