+++ frontend/src/pages/GraphicalFields.tsx
@@ -32,7 +32,7 @@ import CheckCircleOutlineIcon from "@mui/icons-material/CheckCircleOutline";
 import { Group, Layer, Rect, Stage, Text } from "react-konva";
 import type Konva from "konva";
 import type { KonvaEventObject } from "konva/lib/Node";
-import { useHierarchyData, type HierarchyDataState } from "../components/hierarchy/hooks/useHierarchyData";
+import { useHierarchyData } from "../components/hierarchy/hooks/useHierarchyData";
 import { useExpandedState } from "../components/hierarchy/hooks/useExpandedState";
 import {
   layoutAPI,
@@ -61,198 +61,41 @@ import {
   toContentBounds,
   type ContentBounds,
   type PanSession,
-  type ViewportPadding,
   type ViewportState,
   zoomAroundPoint,
 } from "./graphicalViewport";
+import {
+  DEFAULT_STAGE_HEIGHT,
+  FIELD_INNER_BOTTOM_PADDING,
+  FIELD_INNER_OFFSET_X,
+  FIELD_INNER_OFFSET_Y,
+  FIELD_SNAP_THRESHOLD,
+  FIELD_WORLD_BASE_WIDTH,
+  FIELD_WORLD_MAX_WIDTH,
+  FIELD_WORLD_MIN_HEIGHT,
+  FIELD_WORLD_MIN_WIDTH,
+  FIELD_WORLD_PX_PER_METER,
+  FIELD_WORLD_SCALE_FACTOR,
+  LOCATION_FIELD_GAP,
+  LOCATION_LAYOUT_PADDING,
+  MAX_STAGE_HEIGHT,
+  MIN_STAGE_HEIGHT,
+  PAN_FAST_STEP,
+  PAN_STEP,
+  WORKSPACE_MIN_HEIGHT,
+  WORKSPACE_MIN_WIDTH,
+  ZOOM_STEP,
+  getFitViewportPadding,
+  snapToNeighbors,
+  type BedViewModel,
+  type GraphicalFieldsProps,
+  type GuideLine,
+  type InteractionMode,
+  type Point,
+  type RectViewModel,
+  type SelectedElement,
+} from "./graphicalFieldsGeometry";
 
-interface Point {
-  x: number;
-  y: number;
-}
-
-interface SnapSize {
-  width: number;
-  height: number;
-}
-
-interface RectViewModel {
-  id: number;
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-}
-
-interface BedViewModel extends RectViewModel {
-  id: number;
-  name: string;
-  area: number;
-}
-
-interface GuideLine {
-  orientation: "vertical" | "horizontal";
-  value: number;
-  start: number;
-  end: number;
-}
-
-interface SnapResult {
-  x: number;
-  y: number;
-  guides: GuideLine[];
-}
-
-interface SelectedElement {
-  type: "field" | "bed";
-  name: string;
-  area: number | null;
-  locationName: string;
-  parentName?: string;
-}
-
-interface GraphicalFieldsProps {
-  showTitle?: boolean;
-  interactionMode?: InteractionMode;
-  onInteractionModeChange?: (mode: InteractionMode) => void;
-  showModeToggle?: boolean;
-  hierarchyData?: HierarchyDataState;
-}
-
-type InteractionMode = "view" | "edit";
-
-const VIEWPORT_PADDING = 24;
-const VIEWPORT_CONTROL_SAFE_AREA_RIGHT = 120;
-const MOBILE_VIEWPORT_PADDING = 12;
-const MOBILE_VIEWPORT_CONTROL_SAFE_AREA_RIGHT = 8;
-const FIELD_INNER_OFFSET_X = 10;
-const FIELD_LABEL_HEIGHT = 24;
-const FIELD_INNER_OFFSET_Y = FIELD_INNER_OFFSET_X + FIELD_LABEL_HEIGHT;
-const FIELD_INNER_BOTTOM_PADDING = FIELD_INNER_OFFSET_X;
-const SNAP_THRESHOLD = 8;
-const FIELD_SNAP_THRESHOLD = 14;
-const DEFAULT_STAGE_HEIGHT = 420;
-const MIN_STAGE_HEIGHT = 320;
-const MAX_STAGE_HEIGHT = 560;
-const ZOOM_STEP = 1.2;
-const PAN_STEP = 80;
-const PAN_FAST_STEP = 180;
-const WORKSPACE_MIN_WIDTH = 20000;
-const WORKSPACE_MIN_HEIGHT = 20000;
-const LOCATION_LAYOUT_PADDING = 20;
-const LOCATION_FIELD_GAP = 24;
-const FIELD_WORLD_PX_PER_METER = 18;
-const FIELD_WORLD_BASE_WIDTH = 560;
-const FIELD_WORLD_MIN_WIDTH = 560;
-const FIELD_WORLD_MAX_WIDTH = 920;
-const FIELD_WORLD_MIN_HEIGHT = 220;
-const FIELD_WORLD_SCALE_FACTOR = 12;
-
-const getFitViewportPadding = (stageWidth: number): ViewportPadding => ({
-  top: stageWidth < 600 ? MOBILE_VIEWPORT_PADDING : VIEWPORT_PADDING,
-  right:
-    (stageWidth < 600 ? MOBILE_VIEWPORT_PADDING : VIEWPORT_PADDING) +
-    (stageWidth < 600
-      ? MOBILE_VIEWPORT_CONTROL_SAFE_AREA_RIGHT
-      : VIEWPORT_CONTROL_SAFE_AREA_RIGHT),
-  bottom: stageWidth < 600 ? MOBILE_VIEWPORT_PADDING : VIEWPORT_PADDING,
-  left: stageWidth < 600 ? MOBILE_VIEWPORT_PADDING : VIEWPORT_PADDING,
-});
-const snapToNeighbors = (
-  currentId: number,
-  position: Point,
-  size: SnapSize,
-  neighbors: RectViewModel[],
-  threshold: number = SNAP_THRESHOLD,
-): SnapResult => {
-  let snappedX = position.x;
-  let snappedY = position.y;
-  let bestXDelta = threshold + 1;
-  let bestYDelta = threshold + 1;
-  const guides: GuideLine[] = [];
-
-  const currentXPoints = [
-    { value: position.x },
-    { value: position.x + size.width / 2 },
-    { value: position.x + size.width },
-  ];
-  const currentYPoints = [
-    { value: position.y },
-    { value: position.y + size.height / 2 },
-    { value: position.y + size.height },
-  ];
-
-  neighbors
-    .filter((neighbor) => neighbor.id !== currentId)
-    .forEach((neighbor) => {
-      const neighborXPoints = [
-        { value: neighbor.x },
-        { value: neighbor.x + neighbor.width / 2 },
-        { value: neighbor.x + neighbor.width },
-      ];
-      const neighborYPoints = [
-        { value: neighbor.y },
-        { value: neighbor.y + neighbor.height / 2 },
-        { value: neighbor.y + neighbor.height },
-      ];
-
-      currentXPoints.forEach((currentPoint) => {
-        neighborXPoints.forEach((neighborPoint) => {
-          const delta = neighborPoint.value - currentPoint.value;
-          const absDelta = Math.abs(delta);
-          if (absDelta <= threshold && absDelta < bestXDelta) {
-            bestXDelta = absDelta;
-            snappedX = position.x + delta;
-            guides.push({
-              orientation: "vertical",
-              value: neighborPoint.value,
-              start: Math.min(position.y, neighbor.y),
-              end: Math.max(
-                position.y + size.height,
-                neighbor.y + neighbor.height,
-              ),
-            });
-          }
-        });
-      });
-
-      currentYPoints.forEach((currentPoint) => {
-        neighborYPoints.forEach((neighborPoint) => {
-          const delta = neighborPoint.value - currentPoint.value;
-          const absDelta = Math.abs(delta);
-          if (absDelta <= threshold && absDelta < bestYDelta) {
-            bestYDelta = absDelta;
-            snappedY = position.y + delta;
-            guides.push({
-              orientation: "horizontal",
-              value: neighborPoint.value,
-              start: Math.min(position.x, neighbor.x),
-              end: Math.max(
-                position.x + size.width,
-                neighbor.x + neighbor.width,
-              ),
-            });
-          }
-        });
-      });
-    });
-
-  const latestVerticalGuide = [...guides]
-    .reverse()
-    .find((guide) => guide.orientation === "vertical");
-  const latestHorizontalGuide = [...guides]
-    .reverse()
-    .find((guide) => guide.orientation === "horizontal");
-
-  return {
-    x: snappedX,
-    y: snappedY,
-    guides: [
-      ...(latestVerticalGuide ? [latestVerticalGuide] : []),
-      ...(latestHorizontalGuide ? [latestHorizontalGuide] : []),
-    ],
-  };
-};
 
 export default function GraphicalFields({
   showTitle = true,
